`linera-service`: move `assign_new_chain_to_key` to `linera-client`
